lsp: refactor types
Having all these types in the lsp package makes it harder to create
subpackages in the LSP without cyclic import issues.

This PR moves them all under a types subpackage.
