+++ internal/lsp/inlayhint.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/open-policy-agent/opa/ast"
 	"github.com/open-policy-agent/opa/types"
+
+	types2 "github.com/styrainc/regal/internal/lsp/types"
 )
 
 func createInlayTooltip(named *types.NamedType) string {
@@ -15,8 +17,8 @@ func createInlayTooltip(named *types.NamedType) string {
 	return fmt.Sprintf("%s\n\nType: `%s`", named.Descr, named.Type.String())
 }
 
-func getInlayHints(module *ast.Module) []InlayHint {
-	inlayHints := make([]InlayHint, 0)
+func getInlayHints(module *ast.Module) []types2.InlayHint {
+	inlayHints := make([]types2.InlayHint, 0)
 
 	for _, call := range AllBuiltinCalls(module) {
 		for i, arg := range call.Builtin.Decl.NamedFuncArgs().Args {
@@ -27,13 +29,13 @@ func getInlayHints(module *ast.Module) []InlayHint {
 			}
 
 			if named, ok := arg.(*types.NamedType); ok {
-				inlayHints = append(inlayHints, InlayHint{
+				inlayHints = append(inlayHints, types2.InlayHint{
 					Position:     positionFromLocation(call.Args[i].Location),
 					Label:        named.Name + ":",
 					Kind:         2,
 					PaddingLeft:  false,
 					PaddingRight: true,
-					Tooltip: MarkupContent{
+					Tooltip: types2.MarkupContent{
 						Kind:  "markdown",
 						Value: createInlayTooltip(named),
 					},
